add 'build.sh run' to launch app

+++ build.sh
@@ -1,7 +1,11 @@
 cd src
-gem install bundler
-bundle install
-bundle exec rake db:drop
-bundle exec rake db:create
-bundle exec rake db:migrate
+if [ "${1}" == "run" ]; then
+  ruby app.rb
+else
+  gem install bundler
+  bundle install
+  bundle exec rake db:drop
+  bundle exec rake db:create
+  bundle exec rake db:migrate
+fi
